Important fix to get the size of the files correct. Needed as unit reads to the end of the file and we don't want to read a bung of zeros.
Polygons draw (but no colouring or identification) but cause all the lines
to have no information associated with them.  When you hover over there is no 'tooltip' or when you click it just says there is nothing there.

git-svn-id: http://svn.mkgmap.org.uk/mkgmap/trunk@57 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/ImgFile.java
@@ -51,7 +51,7 @@ public abstract class ImgFile {
 		return writer.position();
 	}
 
-	void position(int pos) {
+	void position(long pos) {
 		writer.position(pos);
 	}
